Move async runner tests to dedicated module
